update and delete api changes

+++ src/teachers/teachers.service.ts
@@ -4,10 +4,13 @@ import { Model } from 'mongoose';
 import { Teacher, TeacherDocument } from './schemas/teacher.schema';
 import { CreateTeacherDto } from './dto/create-teacher.dto';
 import { UpdateTeacherDto } from './dto/update-teacher.dto';
+import { User, UserDocument } from 'src/users/schemas/user.schema';
 
 @Injectable()
 export class TeacherService {
-  constructor(@InjectModel(Teacher.name) private teacherModel: Model<TeacherDocument>) {}
+  constructor(@InjectModel(Teacher.name) private teacherModel: Model<TeacherDocument>,
+ @InjectModel(User.name)
+  private userModel: Model<UserDocument>,) {}
 
   async create(userId: string, dto: CreateTeacherDto): Promise<Teacher> {
     const teacher = new this.teacherModel({ ...dto, user: userId });
@@ -24,14 +27,46 @@ export class TeacherService {
     return teacher;
   }
 
-  async update(id: string, dto: UpdateTeacherDto): Promise<Teacher> {
-    const updated = await this.teacherModel.findByIdAndUpdate(id, dto, { new: true });
-    if (!updated) throw new NotFoundException('Teacher not found');
-    return updated;
+async update(id: string, dto: UpdateTeacherDto): Promise<Teacher> {
+
+  if (!dto) {
+    throw new NotFoundException('Update body is missing');
   }
 
-  async remove(id: string): Promise<void> {
-    const deleted = await this.teacherModel.findByIdAndDelete(id);
-    if (!deleted) throw new NotFoundException('Teacher not found');
+  const teacher = await this.teacherModel.findById(id);
+  if (!teacher) {
+    throw new NotFoundException('Teacher not found');
   }
+
+  if (dto?.name || dto?.email) {
+    await this.userModel.findByIdAndUpdate(
+      teacher.user,
+      {
+        ...(dto.name && { name: dto.name }),
+        ...(dto.email && { email: dto.email }),
+      }
+    );
+  }
+
+  const { name, email, ...teacherFields } = dto;
+
+  Object.assign(teacher, teacherFields);
+
+  await teacher.save();
+  await teacher.populate('user');
+
+  return teacher;
+}
+
+  async remove(id: string): Promise<void> {
+
+  const teacher = await this.teacherModel.findById(id);
+  if (!teacher) {
+    throw new NotFoundException('Teacher not found');
+  }
+
+  await this.userModel.findByIdAndDelete(teacher.user);
+
+  await this.teacherModel.findByIdAndDelete(id);
+}
 }
